Allow a client to unsubscribe to channel
This patch allows a client to unsubscribe to a channel. It is used by
example by collectd where we can have different checks. In StackLight we
have a remote collectd that is running and also a collectd daemon. So
there is two collectd processes. By default we subscribe to all channel
according to our roles. But we don't want to subscribe to
collectd.client, only to collectd.remote_client.

Change-Id: I1a6d42abd93b6288740832cfc07fc08ae0559f12