[MMO-282] Adjust Sensu filter config for sfdc
Allow configuring a filter definition for the sfdc handler in the model.

Override the `filter_repeated()` function in sfdc plugin.  There is
little point in having this function filter out repeated alerts since
the destination because the destination is Salesforce.  Repeated alerts
will simply be appened to the same `MOS Alert`.

This also lets us configure filters entirely from the model/Sensu
config, which means we can configure these easier per-site.

As it was, if I wanted to configure the sfdc handler to not create an
alert unless it triggered three time in a row, the `filter_repeated()`
function would mask that third alert.

Change-Id: I43553fb82fcd7b296f14bf1e48e45e2215aaa534